+++ apps/decodex/src/orchestrator/tests/recovery_reconciliation/activity_markers/live_activity.rs
@@ -1,0 +1,3 @@
+mod child_process;
+mod model_execution;
+mod repo_gate;
