+++ apps/decodex/src/orchestrator/tests/runtime_repo_gate/blocking_recovery/retry_resume/cross_attempt_handoff.rs
@@ -1,0 +1,3 @@
+mod active_handoff;
+mod latest_phase;
+mod terminal_finalize;
