+++ apps/decodex/src/research_design/tests/latent_contracts.rs
@@ -1,0 +1,2 @@
+mod feedback;
+mod persistence;
